update existing tests and some renaming
