+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/Users/Application/UseCases/CreateUserIfNotExists/CreateUserIfNotExistsUseCaseTests.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DfE.GIAP.Core.UnitTests.Users.Application.UseCases.CreateUserIfNotExists;
-internal class CreateUserIfNotExistsUseCaseTests
+﻿namespace DfE.GIAP.Core.UnitTests.Users.Application.UseCases.CreateUserIfNotExists;
+public sealed class CreateUserIfNotExistsUseCaseTests
 {
+
 }
