+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/Users/Application/UseCases/UpdateLastLoggedIn/UpdateLastLoggedInUseCaseTests.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DfE.GIAP.Core.UnitTests.Users.Application.UseCases.UpdateLastLoggedIn;
-internal class UpdateLastLoggedInUseCaseTests
+﻿namespace DfE.GIAP.Core.UnitTests.Users.Application.UseCases.UpdateLastLoggedIn;
+public sealed class UpdateLastLoggedInUseCaseTests
 {
+
 }
